+++ src/test/java/org/folio/rest/api/InstanceStorageTest.java
@@ -476,6 +476,36 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
+  public void canSearchByClassificationNumberWithArrayModifier()
+    throws MalformedURLException,
+    InterruptedException,
+    ExecutionException,
+    TimeoutException {
+    
+    createInstancesWithClassificationNumbers();
+
+    JsonObject allInstances = searchForInstances("classifications =/@classificationNumber \"K1 .M385\"");
+
+    assertThat(allInstances.getInteger("totalRecords"), is(1));
+    assertThat(allInstances.getJsonArray("instances").getJsonObject(0).getString("title"), is("Long Way to a Small Angry Planet"));
+  }
+
+  @Test
+  public void canSearchByClassificationNumberWithoutArrayModifier()
+    throws MalformedURLException,
+    InterruptedException,
+    ExecutionException,
+    TimeoutException {
+    
+    createInstancesWithClassificationNumbers();
+
+    JsonObject allInstances = searchForInstances("classifications =\"K1 .M385\"");
+
+    assertThat(allInstances.getInteger("totalRecords"), is(1));
+    assertThat(allInstances.getJsonArray("instances").getJsonObject(0).getString("title"), is("Long Way to a Small Angry Planet"));
+  }
+
+  @Test
   public void canSearchUsingKeywordIndex()
     throws MalformedURLException,
     InterruptedException,
@@ -2652,6 +2682,32 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
       containsInAnyOrder(notSuppressedInstance.getId(), notSuppressedInstanceDefault.getId()));
   }
 
+  private void createInstancesWithClassificationNumbers() throws InterruptedException,
+    ExecutionException, TimeoutException{
+
+    UUID firstInstanceId = UUID.randomUUID();
+    UUID secondInstanceId = UUID.randomUUID();
+    UUID classificationTypeId = UUID.randomUUID();
+
+    JsonObject firstClassifications = new JsonObject();
+    firstClassifications.put("classificationTypeId", classificationTypeId.toString());
+    firstClassifications.put("classificationNumber", "K1 .M385");
+
+    JsonObject secondClassifications = new JsonObject();
+    secondClassifications.put("classificationTypeId", classificationTypeId.toString());
+    secondClassifications.put("classificationNumber", "KB1 .A437");
+
+    JsonObject firstInstanceToCreate = smallAngryPlanet(firstInstanceId)
+    .put("classifications", new JsonArray()
+    .add(firstClassifications));
+    JsonObject secondInstanceToCreate = nod(secondInstanceId)
+    .put("classifications", new JsonArray()
+    .add(secondClassifications));
+
+    createInstance(firstInstanceToCreate);
+    createInstance(secondInstanceToCreate);
+  }
+
   private JsonObject createRequestForMultipleInstances(Integer numberOfInstances) {
     JsonArray instancesArray = new JsonArray();
 
@@ -2777,12 +2833,11 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   private JsonObject nod(UUID id) {
     JsonArray identifiers = new JsonArray();
     identifiers.add(identifier(UUID_ASIN, "B01D1PLMDO"));
-
     JsonArray contributors = new JsonArray();
     contributors.add(contributor(UUID_PERSONAL_NAME, "Barnes, Adrian"));
-
     JsonArray tags = new JsonArray();
     tags.add("test-tag");
+
     return createInstanceRequest(id, "TEST", "Nod",
       identifiers, contributors, UUID_INSTANCE_TYPE, tags);
   }