+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 20.2.0 IN-PROGRESS
 
+* Adds full-text index for classifications field in instance table (MODINVSTOR-716)
 * Adds `effective location` attribute to holdings record and inventory-hierarchy API (MODINVSTOR-669, MODINVSTOR-670) 
 * Defaults instance.previouslyHeld to false (MODINVSTOR-454)
 * Provides `holdings-storage 4.6` (MODINVSTOR-669